Run uwsgi-emperor as root
(As I'm not a paranoid sysadmin. ;-)

From the docs:

The emperor is normally run as root, setting the UID and GID in each
instance’s config. The vassal instance then drops privileges before
serving requests. In this mode, if your users have access to their own
uWSGI configuration files, you can’t trust them to set the correct uid
and gid. You could run the emperor as unprivileged user (with uid and
gid) but all of the vassals would then run under the same user, as
unprivileged users are not able to promote themselves to other users.
For this case the Tyrant mode is available – just add the emperor-tyrant
option.

Source: https://uwsgi-docs.readthedocs.io/en/latest/Emperor.html